requirements.txt
aiosqlite==0.17.0
anyio==3.7.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.96.0
h11==0.14.0
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
pydantic==1.10.8
PyJWT==2.7.0
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.19.3
typing-extensions==4.6.3
uvicorn==0.22.0
